int to float32

+++ pkg/math/math_operations_interfaces.go
@@ -1,5 +1,5 @@
 package pkg
 
 type MathOperations interface {
-	DoMath(x float64, y float64, operationSign string, result *float64) error
+	DoMath(x float32, y float32, operationSign string, result *float32) error
 }
